% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ly.R
\name{generate_ly}
\alias{generate_ly}
\title{Create a .ly file}
\usage{
generate_ly(sample, bars, file_name = "my_song.ly")
}
\arguments{
\item{sample}{a vector of wave objects}

\item{bars}{number of bars}

\item{file_name}{name of the file to export}
}
\description{
Allows you to create a playable .ly file.
}
\examples{
sample <- c(sine(440), sine(220))
generate_ly(sample, 1)
}
