% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rhythm_sample.R
\name{generate_rhythm_sample}
\alias{generate_rhythm_sample}
\title{Create a sample of music notes with rhythm.}
\usage{
generate_rhythm_sample(num_bars)
}
\arguments{
\item{num_bars}{number of bars/measures, in common time signature, that will be produced.}
}
\value{
a list of wave objects
}
\description{
Allows you to create a sample of Wave objects with notes of varying speeds. Where the notes are 
sampled randomly from the notes found on a standard 88 note piano.
}
\examples{
rhythm_sample <- generate_rhythm_sample(5)
generate_lilypond(rhythm_sample, 5)
}
