% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_lilypond.R
\name{generate_lilypond}
\alias{generate_lilypond}
\title{Create a .ly file}
\usage{
generate_lilypond(sample, bars, file_name = "my_song.ly",
  quiet = FALSE)
}
\arguments{
\item{sample}{a vector of wave objects}

\item{bars}{number of bars}

\item{file_name}{name of the file to export}
}
\value{
a file that can be executed by the LilyPond program
}
\description{
Allows you to create a playable .ly file.
}
\examples{
sample <- c(sine(440), sine(220))
generate_ly(sample, 1)
}
